+++ src/index.ts
@@ -1,8 +1,2 @@
-import * as utils from './utils';
-
-export * from './types';
+export * from './core';
 export * from './readers';
-export * from './api';
-export * from './compile';
-
-export { utils };