+++ src/schemas.test.ts
@@ -278,4 +278,94 @@ describe('getConnectors', () => {
       },
     });
   });
+
+  it('always respects local variables before data arguments', () => {
+    const source = `type Translation {
+      key: String
+      content: String
+      language: String
+    }
+
+    type Query {
+      items(language: String): [Translation] {
+        use(get('api/snippet'), res => {
+          const { snippets, languages } = res;
+
+          languages.forEach(language => {
+            snippets[language].map(snippet => ({ ...snippet, language }));
+          });
+        })
+      }
+    }`;
+    const result = getConnectors(source);
+    expect(result).toEqual({
+      Query: {
+        items:
+          "try { const use = ((x, cb) => cb(x)); return use(await $api.get('api/snippet'), ((res) => { const { snippets, languages } = res; languages.forEach(((language) => { snippets[language].map(((snippet) => ({ ...(snippet), language: language }))); })); })); } catch (err) { throw new Error(JSON.stringify(err)); }",
+      },
+    });
+  });
+
+  it('includes array spread correctly', () => {
+    const source = `type Query {
+      items: [Item] {
+        use(get('/api/item'), items => {
+          const arr = [];
+          arr.push(...items);
+          return arr;
+        })
+      }
+    }`;
+    const result = getConnectors(source);
+    expect(result).toEqual({
+      Query: {
+        items:
+          "try { const use = ((x, cb) => cb(x)); return use(await $api.get('/api/item'), ((items) => { const arr = []; arr.push(...items); return arr; })); } catch (err) { throw new Error(JSON.stringify(err)); }",
+      },
+    });
+  });
+
+  it('should respect lazy loading also in conditionals', () => {
+    const source = `type Query {
+      items(arg: String): [Translation] {
+        arg ? get('a').snippets.map(snippet => ({ ...snippet, arg })) : get('b')
+      }
+    }`;
+    const result = getConnectors(source);
+    expect(result).toEqual({
+      Query: {
+        items:
+          "try { return (($data.arg) ? (await (async () => { const _0 = await $api.get('a'); return _0.snippets.map(((snippet) => ({ ...(snippet), arg: $data.arg }))); })()) : (await $api.get('b'))); } catch (err) { throw new Error(JSON.stringify(err)); }",
+      },
+    });
+  });
+
+  it('should handle all kinds of conditionals', () => {
+    const source = `type Query {
+      test(id: ID): Int {
+        use(get('api/item').items, ([item]) => {
+          item = item ? item : post('api/item', {});
+          id = id ? id : use(get('api/foo')
+            .items.filter(m => m.type === 'bar'), ([foo]) => {
+              foo = foo ? foo : post('api/foo', {
+                name: 'Foo',
+                type: 'bar',
+              });
+              return foo.id;
+            });
+
+          return post(\`api/item/\${id}\`, {
+            target: item.id,
+          });
+        })
+      }
+    }`;
+    const result = getConnectors(source);
+    expect(result).toEqual({
+      Query: {
+        test:
+          "try { const use = ((x, cb) => cb(x)); const _0 = await $api.get('api/item'); return await use(_0.items, (async ([item]) => { (item = ((item) ? (item) : (await $api.post('api/item', ({  }))))); ($data.id = (($data.id) ? ($data.id) : (await (async () => { const _1 = await $api.get('api/foo'); return await use(_1.items.filter(((m) => (m.type === 'bar'))), (async ([foo]) => { (foo = ((foo) ? (foo) : (await $api.post('api/foo', ({ name: 'Foo', type: 'bar' }))))); return foo.id; })); })()))); return await $api.post(`api/item/${$data.id}`, ({ target: item.id })); })); } catch (err) { throw new Error(JSON.stringify(err)); }",
+      },
+    });
+  });
 });